:art: :newspaper: :tractor: Restructurize modules
