+++ src/Importify/Paths.hs
@@ -2,14 +2,18 @@
 
 -- | This module contains common utilities for working with importify cache.
 
-module Importify.Cache
+module Importify.Paths
        ( -- * Predefined directories
          cacheDir
        , cachePath
+       , extensionsFile
+       , extensionsPath
        , modulesFile
        , modulesPath
        , symbolsDir
        , symbolsPath
+       , targetsFile
+       , targetsPath
 
          -- * Utility functions to work with directories
        , guessCabalName
@@ -32,10 +36,21 @@ symbolsPath = [reldir|symbols/|]
 modulesPath :: Path Rel File
 modulesPath = [relfile|modules|]
 
-cacheDir, modulesFile, symbolsDir :: FilePath
+-- | Path to JSON-encoded Map from project module name to
+-- its target (i.e. __library__, __executable__).
+targetsPath :: Path Rel File
+targetsPath = [relfile|targets|]
+
+-- | Path to JSON-encoded Map from target to its list of default extensions.
+extensionsPath :: Path Rel File
+extensionsPath = [relfile|extensions|]
+
+cacheDir, extensionsFile, modulesFile, symbolsDir, targetsFile :: FilePath
 cacheDir    = fromRelDir  cachePath
+extensionsFile = fromRelFile extensionsPath
 modulesFile = fromRelFile modulesPath
 symbolsDir  = fromRelDir  symbolsPath
+targetsFile = fromRelFile targetsPath
 
 -- TODO: probably not reliable, instead file with
 -- .cabal extension should be searched
